+++ pizza_time/manage.php
@@ -138,7 +138,7 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST' && isset($_POST['clearTables'])) {
                         <a class="nav-link text-white fw-bold px-lg-5" href="manage.php">Managment</a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link text-white fw-bold px-lg-5" href="#">Orders</a>
+                        <a class="nav-link text-white fw-bold px-lg-5" href="order.php">Orders</a>
                     </li>
                     <li class="nav-item">
                         <a class="nav-link text-white fw-bold px-lg-5" href="server_lookup.php">Server_lookup</a>
