+++ pizza_time/server_lookup.php
@@ -73,7 +73,7 @@
                         <a class="nav-link text-white fw-bold px-lg-5" href="manage.php">Managment</a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link text-white fw-bold px-lg-5" href="#">Orders</a>
+                        <a class="nav-link text-white fw-bold px-lg-5" href="order.php">Orders</a>
                     </li>
                     <li class="nav-item">
                         <a class="nav-link text-white fw-bold px-lg-5" href="server_lookup.php">Server_lookup</a>
